+++ src/Aggregation/CumulativeSum.php
@@ -1,0 +1,56 @@
+<?php
+
+namespace Elastica\Aggregation;
+
+use Elastica\Exception\InvalidException;
+
+/**
+ * Class CumulativeSum.
+ *
+ * @see https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-pipeline-cumulative-sum-aggregation.html
+ */
+class CumulativeSum extends AbstractAggregation
+{
+    public const DEFAULT_FORMAT_VALUE = null;
+
+    public function __construct(string $name, ?string $bucketsPath = null)
+    {
+        parent::__construct($name);
+
+        if (null !== $bucketsPath) {
+            $this->setBucketsPath($bucketsPath);
+        }
+    }
+
+    /**
+     * Set the buckets_path for this aggregation.
+     *
+     * @return $this
+     */
+    public function setBucketsPath(string $bucketsPath): self
+    {
+        return $this->setParam('buckets_path', $bucketsPath);
+    }
+
+    /**
+     * Set the format for this aggregation.
+     *
+     * @return $this
+     */
+    public function setFormat(?string $format = null): self
+    {
+        return $this->setParam('format', $format);
+    }
+
+    /**
+     * @throws InvalidException If buckets path is not set
+     */
+    public function toArray(): array
+    {
+        if (!$this->hasParam('buckets_path')) {
+            throw new InvalidException('Buckets path is required');
+        }
+
+        return parent::toArray();
+    }
+}